working better visualization
